requirements.txt
aiogram
qrcode
pillow
python-dotenv
opencv-python